Producer/requirements.txt
requests==2.26.0
kafka_python==2.0.2
beautifulsoup4==4.10.0
